census: Add action buttons for each voter
This patch adds actions buttons for each row in the census table.
Buttons added are Activate, Deactivate, Remove and Send auth codes.

These buttons select the row and then call the same action as actions
dropdown, all other rows are deselected so this should work similar to
select only one row and click in the actions dropdown.

Buttons are hidden by default and showed on hover the row.

I've add translations for these actions but not for gl and ca.