census: Comment for activate/deactivate actions
This patch allows comments for activate and deactivate actions. A
textarea is available in the activate action and the content of that
will go as comment.

To show the comment in the activity log I've to change a little the
code because the angular template always resolve the {{obj.metadata.comment}}
as undefined, I don't know why, so I've added other attr to obj called
metadatacomment.