Add cumulative voting support
This patch adds the cumulative voting tally type.

The cumulative tally_type will only be supported by the
simultaneous questions layout, so this patch ensure that the user
can only select the correct layout/tally_type combination and
shows a descriptive message for the user.

A new extra_option is also added so the
cumulative_number_of_questions could be configured in the admin
interface.